Stripe Setup

+++ CodersGear.Models/ShoppingCart.cs
@@ -17,12 +17,12 @@ namespace CodersGear.Models
         [Range(1, 1000, ErrorMessage = "Count must be at least 1")]
         public int Count { get; set; }
         [NotMapped]
-        public double Price { get; set; }
+        public decimal Price { get; set; }
         public string ApplicationUserId { get; set; }
         [ForeignKey("ApplicationUserId")]
         [ValidateNever]
         public ApplicationUser ApplicationUser { get; set; }
-        public double OrderTotal { get; set; }
+        public decimal OrderTotal { get; set; }
